Add separate WAGTAILLOCALIZE_LANGUAGES setting
